fast commit = )

+++ project/PPInterpreter/tokenstream.h
@@ -29,7 +29,7 @@ class TokenStream {
         return tokens_[current_pos_].value_ == value;
     }
     bool End() {
-        return current_pos_ == static_cast<ssize_t>(tokens_.size() - 1);
+        return current_pos_ == static_cast<int>(tokens_.size() - 1);
     }
     bool CompareTypeWithRollback(TokenType type);
     bool CompareValueWithRollback(string value);
@@ -38,7 +38,7 @@ class TokenStream {
     DISABLE_COPY_AND_ASSIGN(TokenStream);
 
     vector<Token> tokens_;
-    ssize_t current_pos_;
+    int current_pos_;
 };
 
 #endif // TOKENSTREAM_H
